+++ laravel5.4/app/Admin/Models/Admin.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Admin\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use DB;
