+++ laravel5.4/app/Admin/Models/AdminRole.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace App\Admin\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class CatType extends Model
+class AdminRole extends Model
 {
-    protected $table = 'cat_type';  //表名
+    protected $table = 'admin_role';  //表名
     public $timestamps = false;  //过滤默认的字段
 }
