+++ laravel5.4/app/Admin/Models/CatType.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace App\Admin\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class AdminRole extends Model
+class CatType extends Model
 {
-    protected $table = 'admin_role';  //表名
+    protected $table = 'cat_type';  //表名
     public $timestamps = false;  //过滤默认的字段
 }
