+++ laravel5.4/app/Admin/Models/GradeType.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace App\Admin\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class SubjectType extends Model
+class GradeType extends Model
 {
-    protected $table = 'subject_type';  //表名
+    protected $table = 'grade_type';  //表名
     public $timestamps = false;  //过滤默认的字段
 }
